refactor(types): consolidate scattered Protocol classes into aminx.types
Completes the aminx.types/ consolidation convention already established by
types/protocols.py and types/stages.py: moves every Protocol class defined
outside types/ (except the two explicitly reference-only tiling/iterator.py
classes gated by EPIC #1541's T2.GATE parity gate) into the appropriate
types/ submodule, and updates every import site to pull from aminx.types.

Moves (old -> new):
- BatchLogitFn, TieGroupFuseFn: inference/logits.py -> types/stages.py
  (this also resolves a pre-existing TYPE_CHECKING-guarded circular-import
  workaround in stages.py -- the two names now live where they're consumed)
- CalibrationModule: potts/calibration.py -> types/protocols.py
- ConformationalStates: run/specs.py -> types/protocols.py (previously a
  TYPE_CHECKING-local class; empirically the 5th file matching the task's
  "FIVE files define Protocol outside types/" scope)
- ScanTransition, DedupFn, GatherFn: tiling/strategy.py -> new
  types/tiling_protocols.py (batch-planning internals, distinct domain from
  types/protocols.py's model/pipeline protocols)
- DistributedLockBackend: host/campaign.py -> new types/host_protocols.py
  (host infra interface, distinct from callable-shaped protocols)

All 8 names re-exported from types/__init__.py's flat namespace. Import
sites updated in tiling/__init__.py and tests/tiling/test_strategy.py.
TYPE_CHECKING-guarded where usage is annotation-only (host/campaign.py,
tiling/strategy.py); direct top-level where a real runtime value is needed
(inference/logits.py's Registry[type[BatchLogitFn]]).

tiling/iterator.py (MapIterator/ScanIterator) and tiling/dispatch.py are
untouched per explicit scope exclusion.

Incidental: host/plan.py picks up a single-line ruff --fix (redundant blank
line before an import block) surfaced while running the required whole-tree
ruff check gate; unrelated to the Protocol moves.

Verified: uv run ty check src/ (223 baseline diagnostics, +1 pre-existing/
unrelated diagnostic in training/trainer.py surfaced by a `uv sync` cache
refresh, file untouched), uv run ruff check src/ (592 baseline remaining ->
590, net improvement), uv run pytest -x -q (1401 passed, 14 skipped, 79
deselected, 1 xpassed, 0 failed).
