+++ src/aminx/tiling/strategy.py
@@ -13,39 +13,17 @@ from __future__ import annotations
 
 import dataclasses
 from dataclasses import dataclass
-from typing import Any, Generic, Protocol, TypeVar, runtime_checkable
+from typing import TYPE_CHECKING, Any, Generic, TypeVar
 
 import jax
 import numpy as np
 
+if TYPE_CHECKING:
+  from aminx.types.tiling_protocols import DedupFn, GatherFn, ScanTransition
+
 C = TypeVar("C")  # carry type
 X = TypeVar("X")  # per-step input
 Y = TypeVar("Y")  # per-step output
-
-
-@runtime_checkable
-class ScanTransition(Protocol, Generic[C, X, Y]):
-  """Transition function for carry-bearing axis scan.
-
-  Must satisfy: (carry, x) -> (carry, y)
-  where carry has fixed shape across all iterations of the axis.
-  """
-
-  def __call__(self, carry: C, x: X) -> tuple[C, Y]: ...
-
-
-@runtime_checkable
-class DedupFn(Protocol):
-  """In-trace static gather: select K_bucket unique elements from xs."""
-
-  def __call__(self, xs: Any, unique_indices: Any) -> Any: ...
-
-
-@runtime_checkable
-class GatherFn(Protocol):
-  """In-trace scatter: expand K_bucket unique results back to N positions."""
-
-  def __call__(self, ys_unique: Any, index_map: Any) -> Any: ...
 
 
 def _default_dedup_fn(xs: Any, unique_indices: Any) -> Any:
@@ -121,4 +99,4 @@ class DedupGather:
 # Sealed union — all four are concrete and exhaustive.
 AxisStrategy = Vmap | SafeMap | Scan | DedupGather
 
-__all__ = ["AxisStrategy", "DedupFn", "DedupGather", "GatherFn", "SafeMap", "Scan", "ScanTransition", "Vmap"]
+__all__ = ["AxisStrategy", "DedupGather", "SafeMap", "Scan", "Vmap"]
